requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pyvisa==1.14.1
grpcio==1.59.0
grpcio-tools==1.59.0
plotly==5.17.0
pydantic==2.5.0
